refactor: diff_engine.rs を責務別の3サブモジュールに分割 (#18)
- table_diff.rs: テーブル・フィールド差分（5テスト）
- layout_diff.rs: レイアウト・オブジェクト差分（3テスト）
- catalog_diff.rs: テーブルオカレンス・リレーション差分（9テスト）
- mod.rs: 公開型・diff_ddr()・共有ユーティリティ・統合テスト（4テスト）
